// HealthCare.Web/vhc/app/shared/models/meeting-details.model.ts
export class MeetingDetails {
    ID: string;
    OnLineMeetingId: string;
    OnlineMeetingUri: string;
    OnlineMeetingUrl: string;
    Organizer: string;
    Attendees: string;
    JoinUrl: string;
    PatientName: string;
    Status: string;
    AppointmentDate: string;
    QuestionnaireRequired: string;
    QuestionCategory: string;
    StartTime: Date;
    EndTime: Date;
    DoctorsName: string;
    IsDevicesTestDone: boolean;
}